Refactored code to support AppCompatActivity instead of ActionbarActivity.
