+++ app/src/main/java/com/github/mobile/ui/FragmentStatePagerAdapter.java
@@ -16,7 +16,7 @@
 package com.github.mobile.ui;
 
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v4.app.FragmentActivity;
 import android.view.ViewGroup;
 
 /**
@@ -26,14 +26,14 @@ public abstract class FragmentStatePagerAdapter extends
     android.support.v4.app.FragmentStatePagerAdapter implements
     FragmentProvider {
 
-    private final ActionBarActivity activity;
+    private final FragmentActivity activity;
 
     private Fragment selected;
 
     /**
      * @param activity
      */
-    public FragmentStatePagerAdapter(final ActionBarActivity activity) {
+    public FragmentStatePagerAdapter(final FragmentActivity activity) {
         super(activity.getSupportFragmentManager());
 
         this.activity = activity;
@@ -45,7 +45,7 @@ public abstract class FragmentStatePagerAdapter extends
     public FragmentStatePagerAdapter(final Fragment fragment) {
         super(fragment.getChildFragmentManager());
 
-        this.activity = (ActionBarActivity) fragment.getActivity();
+        this.activity = fragment.getActivity();
     }
 
     @Override
