+++ app/src/main/java/com/github/mobile/ui/comment/CommentPreviewPagerAdapter.java
@@ -17,7 +17,7 @@ package com.github.mobile.ui.comment;
 
 import android.content.Context;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v4.app.FragmentActivity;
 
 import com.github.mobile.R;
 import com.github.mobile.ui.FragmentPagerAdapter;
@@ -46,7 +46,7 @@ public class CommentPreviewPagerAdapter extends FragmentPagerAdapter {
      * @param activity
      * @param repo
      */
-    public CommentPreviewPagerAdapter(ActionBarActivity activity,
+    public CommentPreviewPagerAdapter(FragmentActivity activity,
             IRepositoryIdProvider repo) {
         super(activity);
         this.context = activity.getApplicationContext();
