+++ app/src/main/java/com/github/mobile/ui/commit/CommitPagerAdapter.java
@@ -19,6 +19,7 @@ import static com.github.mobile.Intents.EXTRA_BASE;
 import static com.github.mobile.Intents.EXTRA_REPOSITORY;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentActivity;
 import android.support.v7.app.ActionBarActivity;
 
 import com.github.mobile.ui.FragmentStatePagerAdapter;
@@ -39,7 +40,7 @@ public class CommitPagerAdapter extends FragmentStatePagerAdapter {
      * @param repository
      * @param ids
      */
-    public CommitPagerAdapter(ActionBarActivity activity,
+    public CommitPagerAdapter(FragmentActivity activity,
             Repository repository, CharSequence[] ids) {
         super(activity);
