+++ app/src/main/java/com/github/mobile/ui/gist/GistFilesPagerAdapter.java
@@ -18,7 +18,7 @@ package com.github.mobile.ui.gist;
 import static com.github.mobile.Intents.EXTRA_GIST_FILE;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v4.app.FragmentActivity;
 
 import com.github.mobile.ui.FragmentPagerAdapter;
 
@@ -38,7 +38,7 @@ public class GistFilesPagerAdapter extends FragmentPagerAdapter {
      * @param activity
      * @param gist
      */
-    public GistFilesPagerAdapter(ActionBarActivity activity, Gist gist) {
+    public GistFilesPagerAdapter(FragmentActivity activity, Gist gist) {
         super(activity);
 
         Map<String, GistFile> gistFiles = gist.getFiles();
