+++ app/src/main/java/com/github/mobile/ui/issue/IssuesPagerAdapter.java
@@ -23,7 +23,7 @@ import static com.github.mobile.Intents.EXTRA_REPOSITORY_OWNER;
 import static com.github.mobile.Intents.EXTRA_USER;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v4.app.FragmentActivity;
 import android.util.SparseArray;
 import android.view.ViewGroup;
 
@@ -64,7 +64,7 @@ public class IssuesPagerAdapter extends FragmentStatePagerAdapter {
      * @param issueStore
      * @param collaborator
      */
-    public IssuesPagerAdapter(ActionBarActivity activity,
+    public IssuesPagerAdapter(FragmentActivity activity,
             List<RepositoryId> repoIds, int[] issueNumbers,
             IssueStore issueStore, boolean collaborator, boolean owner) {
         super(activity);
@@ -83,7 +83,7 @@ public class IssuesPagerAdapter extends FragmentStatePagerAdapter {
      * @param issueNumbers
      * @param collaborator
      */
-    public IssuesPagerAdapter(ActionBarActivity activity,
+    public IssuesPagerAdapter(FragmentActivity activity,
             Repository repository, int[] issueNumbers,
             boolean collaborator, boolean owner) {
         super(activity);
