+++ app/src/main/java/com/github/mobile/ui/repo/RepositoryPagerAdapter.java
@@ -18,7 +18,7 @@ package com.github.mobile.ui.repo;
 import android.content.res.Resources;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v4.app.FragmentActivity;
 
 import com.github.mobile.R;
 import com.github.mobile.ui.FragmentPagerAdapter;
@@ -55,7 +55,7 @@ public class RepositoryPagerAdapter extends FragmentPagerAdapter {
      * @param activity
      * @param hasIssues
      */
-    public RepositoryPagerAdapter(ActionBarActivity activity,
+    public RepositoryPagerAdapter(FragmentActivity activity,
             boolean hasIssues) {
         super(activity);
