+++ app/src/main/java/com/github/mobile/ui/search/SearchPagerAdapter.java
@@ -17,7 +17,7 @@ package com.github.mobile.ui.search;
 
 import android.content.res.Resources;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v4.app.FragmentActivity;
 
 import com.github.mobile.R;
 import com.github.mobile.ui.FragmentPagerAdapter;
@@ -34,7 +34,7 @@ public class SearchPagerAdapter extends FragmentPagerAdapter {
      *
      * @param activity
      */
-    public SearchPagerAdapter(ActionBarActivity activity) {
+    public SearchPagerAdapter(FragmentActivity activity) {
         super(activity);
 
         resources = activity.getResources();
