+++ app/src/main/java/com/github/mobile/ui/user/UserPagerAdapter.java
@@ -17,7 +17,7 @@ package com.github.mobile.ui.user;
 
 import android.content.res.Resources;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v4.app.FragmentActivity;
 
 import com.github.mobile.R;
 import com.github.mobile.ui.FragmentPagerAdapter;
@@ -33,7 +33,7 @@ public class UserPagerAdapter extends FragmentPagerAdapter {
     /**
      * @param activity
      */
-    public UserPagerAdapter(final ActionBarActivity activity) {
+    public UserPagerAdapter(final FragmentActivity activity) {
         super(activity);
 
         resources = activity.getResources();
